Move trailing whitespace to _after_ closing tags
This is an "issue" with markdown - `**a **` is not properly rendered, whereas
`**a** ` is.

So, for example,  in cases when the final few chars in a bold block are spaces,
remove them and then re-add _after_ we close off the bold block.

We do this with all the inline styles 🙃