Fix bugginess when converting draftjs trailing whitespace to markdown
There was some code in place for this but it wasn't working in all
cases. For inline styles like italic and bold, markdown doesn't play
nice when the final few characters are whitepsace, so this checks for
that and cuts it out, adds the closing markdown character, and _then_
re-adds the trailing whitespace.