Fix markdown stripping blank paragraphs/newlines
This is an option that can be turned on if you want 1:1 parity with
draftjs as far as newlines go.

By default, markdown's rules state that multiple newlines are collapsed.

With this option turned on, when converting back and forth, they will NOT be.

However, it's worth noting that if you then decide to pass the  markdown
to a 3rd party to render it into HTML, unless that 3rd party's parser
has a similar option, the newlines will be collapsed at that point.