Fix issue with surrogate pairs messing up links at EOL
The closing code for entity items was sometimes not run at the end of a block if the block text included emoji made up of multiple characters.  In cases such as this. string.length !== Array.from(string).length, the latter being a more accurate representation of how many characters there actually are.