Clarify documentation about the collapsing of multiple empty lines
I found the documentation ambiguous because it referred to "newlines" which could be either an empty line in the text or the newline character, and one empty line requires two newline characters.  I have slightly adjusted it to be more clear about what it is trying to say.

Potentially closes out https://github.com/Rosey/markdown-draft-js/issues/142